# docs/phase-0/06-api-spec.yaml
...
openapi: 3.0.3
info:
  title: Nogoolin API
  description: |
    REST API for Nogoolin — Premium Religious Product Catalog Platform.

    Layered architecture: Controller → Service → Repository.
    Base path: /api/v1/

    ## Changelog

    | Version | Date | Type | Description |
    |---|---|---|---|
    | 1.0.0 | June 2026 | MAJOR | Initial version. Covers public catalog (categories, products, multi-script search FR-PUB-014), inquiries, admin products/categories/media/inquiries/settings/audit, and future order endpoints (W priority, gated by `delivery_enabled`). |
  version: 1.0.0
  contact:
    name: Tengis (Solo Developer)
servers:
  - url: https://api.nogoolin.mn/api/v1
    description: Production (Railway)
  - url: http://localhost:3001/api/v1
    description: Local development
...
security:
  - {}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT access token (15 min expiry, FR-AUTH-005, NFR-SEC-013)
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 12
  schemas:
...
    ProductStatus:
      type: string
      enum: [draft, published, archived]
    StockStatus:
      type: string
      enum: [in_stock, out_of_stock, pre_order]
    InquiryStatus:
      type: string
      enum: [new, contacted, closed]
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Бурхан зурагнууд"
        slug:
          type: string
          example: "buddha-statues"
        description:
          type: string
          nullable: true
        sort_order:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        image_url:
          type: string
          format: uri
        alt_text:
          type: string
          nullable: true
        sort_order:
          type: integer
        created_at:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/Category'
        name:
          type: string
          example: "Ногоон Дарь Эх"
        name_en:
          type: string
          nullable: true
          example: "Green Tara"
          description: "FR-PUB-013, v1.2.0"
        slug:
          type: string
          example: "nogoon-dar-eh"
        price:
          type: number
          format: decimal
        short_description:
          type: string
          nullable: true
        full_description:
          type: string
          nullable: true
        usage_instruction:
          type: string
          nullable: true
          description: Markdown content (FR-PROD-010)
        status:
          $ref: '#/components/schemas/ProductStatus'
        stock_status:
          $ref: '#/components/schemas/StockStatus'
        is_featured:
          type: boolean
        model_3d_url:
          type: string
          format: uri
          nullable: true
          description: "GLB URL for 360° viewer (FR-PUB-011)"
        search_tags:
          type: array
          items:
            type: string
          nullable: true
          example: ["nogoon dar eh", "green tara", "tara statue"]
          description: "FR-PUB-013, v1.2.0"
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Inquiry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
          nullable: true
        product:
          $ref: '#/components/schemas/Product'
        customer_name:
          type: string
        phone:
          type: string
        message:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/InquiryStatus'
        created_at:
          type: string
          format: date-time
    InquiryInput:
      type: object
      required: [customer_name, phone]
      properties:
        customer_name:
          type: string
          minLength: 2
        phone:
          type: string
          pattern: '^[0-9+\s-]{8,15}$'
        message:
          type: string
        product_id:
          type: string
          format: uuid
          nullable: true
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Order placement is currently disabled"
        code:
          type: string
          example: "DELIVERY_DISABLED"
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request body failed Zod validation (NFR-SEC-003)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid JWT (NFR-SEC-009)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not admin role (FR-AUTH-009)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /inquiries:
    post:
      tags: [Inquiries]
      summary: Submit a product inquiry
      description: |
        Allows a guest (unauthenticated) to submit an inquiry about a
        product, or a general contact inquiry (`product_id` omitted).

        Validation occurs on both frontend and backend (NFR-SEC-003).

        Rate limited to **3 submissions per IP per hour** (FR-INQ-007).
        Saved with status `new` by default (FR-INQ-003).

        Ref: UC-G-007, FR-INQ-001 to FR-INQ-007, SEQ-002
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InquiryInput'
            example:
              customer_name: "Бат-Эрдэнэ"
              phone: "99112233"
              message: "Энэ бүтээгдэхүүн одоо байгаа эсэх, хэрхэн авах вэ?"
              product_id: "8f14e45f-ceea-4c4a-b0b1-000000000010"
      responses:
        '201':
          description: Inquiry created successfully with status `new`
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Inquiry'
              example:
                data:
                  id: "8f14e45f-ceea-4c4a-b0b1-000000000201"
                  product_id: "8f14e45f-ceea-4c4a-b0b1-000000000010"
                  customer_name: "Бат-Эрдэнэ"
                  phone: "99112233"
                  message: "Энэ бүтээгдэхүүн одоо байгаа эсэх, хэрхэн авах вэ?"
                  status: new
                  created_at: "2026-06-13T10:00:00Z"
        '400':
          description: |
            Validation failure — required fields empty or phone number
            format invalid (UC-G-007 Alternative Flow — Validation Failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: |
            Rate limit exceeded — more than 3 submissions from the same IP
            within the past hour (UC-G-007 Exception Flow)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Too many submissions. Please try again later."
                code: "RATE_LIMIT_EXCEEDED"

  # ════════════════════════════════════════════════
  # ADMIN: PRODUCTS
  # ════════════════════════════════════════════════
...
  /admin/inquiries:
    get:
      tags: [Admin - Inquiries]
      summary: List all inquiries
      description: |
        Returns all inquiries, sorted by newest first.
        Supports filtering by status.

        Ref: UC-ADM-009, FR-INQ-004, FR-INQ-006
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by inquiry status (FR-INQ-006)
          schema:
            $ref: '#/components/schemas/InquiryStatus'
      responses:
        '200':
          description: Paginated inquiry list, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inquiry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "8f14e45f-ceea-4c4a-b0b1-000000000201"
                    product_id: "8f14e45f-ceea-4c4a-b0b1-000000000010"
                    product:
                      id: "8f14e45f-ceea-4c4a-b0b1-000000000010"
                      name: "Ногоон Дарь Эх"
                      slug: "nogoon-dar-eh"
                    customer_name: "Бат-Эрдэнэ"
                    phone: "99112233"
                    message: "Энэ бүтээгдэхүүн одоо байгаа эсэх?"
                    status: new
                    created_at: "2026-06-13T10:00:00Z"
                meta:
                  page: 1
                  limit: 12
                  total: 1
                  total_pages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/inquiries/{id}/status:
    patch:
      tags: [Admin - Inquiries]
      summary: Update inquiry status
      description: |
        Updates inquiry status through the lifecycle:
        `new` → `contacted` → `closed`.

        Writes an audit log entry on success.

        Ref: UC-ADM-010, FR-INQ-005
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/InquiryStatus'
            example:
              status: contacted
      responses:
        '200':
          description: Inquiry status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Inquiry'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════
  # ADMIN: SETTINGS (DELIVERY TOGGLE)
  # ════════════════════════════════════════════════
...